+++ picard/webservice/__init__.py
@@ -64,6 +64,7 @@ from picard.util import (
 )
 from picard.util.xml import parse_xml
 from picard.webservice import ratecontrol
+from picard.webservice.utils import port_from_qurl
 
 
 COUNT_REQUESTS_DELAY_MS = 250
@@ -89,42 +90,6 @@ class UnknownResponseParserError(Exception):
     def __init__(self, response_type):
         message = "Unknown parser for response type '%s'. Parser for given response type does not exist." % response_type
         super().__init__(message)
-
-
-def port_from_qurl(qurl):
-    """Returns QUrl port or default ports (443 for https, 80 for http)"""
-    if qurl.scheme() == 'https':
-        return qurl.port(443)
-    return qurl.port(80)
-
-
-def hostkey_from_url(url):
-    """Returns (host, port) from passed url (as string or QUrl)"""
-    if not isinstance(url, QUrl):
-        url = QUrl(url)
-    return (url.host(), port_from_qurl(url))
-
-
-def host_port_to_url(host, port, path=None, scheme=None, as_string=False):
-    """Convert host & port (with optional path and scheme) to an URL"""
-    url = QUrl()
-    if scheme is None:
-        if port == 443:
-            scheme = 'https'
-        else:
-            scheme = 'http'
-    url.setScheme(scheme)
-
-    if ((scheme == 'https' and port != 443)
-            or (scheme == 'http' and port != 80)):
-        url.setPort(port)
-
-    url.setHost(host)
-
-    if path is not None:
-        url.setPath(path)
-
-    return url.toString() if as_string else url
 
 
 class WSRequest(QNetworkRequest):
